// frontend/src/modules/auth/auth.schema.ts
import { z } from "zod";

export const registerSchema = z.object({
  companyName: z.string().min(2, "Company name is required"),
  name: z.string().min(2, "Full name is required"),
  email: z.email("Invalid email address"),
  password: z.string().min(6, "Password must be at least 6 characters"),
});

export const loginSchema = z.object({
  email: z.email("Invalid email address"),
  password: z.string().min(1, "Password is required"),
});